server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic==1.14.0",
    "asyncpg>=0.30.0",
    "fastapi>=0.115.6",
    "greenlet>=3.1.1",
    "pydantic-settings>=2.7.1",
    "sqlmodel>=0.0.22",
    "uvicorn>=0.34.0",
]


[tool.ruff]
line-length = 79
